refactor: shell.py
